+++ pvzclass/PVZ.cpp
@@ -146,9 +146,4 @@ PVZ::Board PVZ::GetBoard()
 	return(address == 0 ? INVALID_BASEADDRESS : Board(address));
 }
 
-PVZ::SaveData PVZ::GetSaveData()
-{
-	return SaveData(Memory::ReadPointer(0x6A9EC0, 0x82C));
-}
-
 #pragma endregion
